operator: policy: operator cleanup (minor)
This minor change drops the "validator" name from the network policy
operator cell. (It doesn't make sense; this cell does a lot more than
policy validation). It moves some files around to make it a bit easier
to read.

It also moves the label validation out of the policyderivative
controller, which is going away, in to the general policy cell.

No functional changes.
